crew_ai/risk_management_agent/pyproject.toml
[project]
name = "risk_managment_agent"
version = "1.0.0"
description = "Automotive Supply Chain Risk Management System using CrewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[google-genai,tools]>=0.86.0",
    "newscatcher-catchall-sdk>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
run_crew = "risk_managment_agent.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[tool.hatch.build.targets.wheel]
packages = ["src/risk_managment_agent"]
